+++ django_scrubber/management/commands/scrub_validation.py
@@ -2,6 +2,7 @@ import sys
 
 from django.core.management.base import BaseCommand
 
+from django_scrubber import settings_with_fallback
 from django_scrubber.services.validator import ScrubberValidatorService
 
 
@@ -13,8 +14,17 @@ class Command(BaseCommand):
 
         found_models = 0
         found_fields = 0
+
+        whitelisted_models = settings_with_fallback('SCRUBBER_VALIDATION_WHITELIST')
+
         if len(non_scrubbed_field_list):
             for model_path, affected_field_list in non_scrubbed_field_list.items():
+
+                if model_path in whitelisted_models:
+                    print(f'Model {model_path!r} was excluded via \'SCRUBBER_VALIDATION_WHITELIST\' and will '
+                          'not be validated.')
+                    continue
+
                 print(f'Model {model_path!r}:')
                 found_models += 1
                 for field in affected_field_list: